add test config file and move all constants there
