+++ app/client/src/pages/Templates/constants.ts
@@ -1,5 +1,3 @@
-import type { Template } from "api/TemplatesApi";
-
 // These will be eventually from an api
 export const functions = [
   {
@@ -42,28 +40,3 @@ export const useCases = [
 export const TEMPLATE_BUILDING_BLOCKS_FILTER_FUNCTION_VALUE = "Building Blocks";
 export const TEMPLATE_ALL_FILTER_FUNCTION_VALUE = "All";
 export const BUILDING_BLOCK_THUMBNAIL_ALT_TEXT = "Building Block Thumbnail";
-
-// Unit Tests
-export const MOCK_BUILDING_BLOCK_TITLE = "Test Building Block";
-export const MOCK_BUILDING_BLOCK_DESCRIPTION =
-  "Description of the test building block";
-export const MOCK_BUILDING_BLOCK_ID = "mockId";
-
-export const mockBuildingBlock: Template = {
-  id: MOCK_BUILDING_BLOCK_ID,
-  userPermissions: ["read", "write"],
-  title: MOCK_BUILDING_BLOCK_TITLE,
-  description: MOCK_BUILDING_BLOCK_DESCRIPTION,
-  appUrl: "https://mockapp.com",
-  gifUrl: "https://mockapp.com/mock.gif",
-  screenshotUrls: [
-    "https://mockapp.com/screenshot1.jpg",
-    "https://mockapp.com/screenshot2.jpg",
-  ],
-  widgets: [],
-  functions: ["Function1", "Function2"],
-  useCases: ["UseCase1", "UseCase2"],
-  datasources: ["Datasource1", "Datasource2"],
-  pages: [],
-  allowPageImport: true,
-};
